Reorganized code a bit. Generators now have submodules
generators.basic -> web3-based interface and CLI
generators.brownie -> brownie (https://github.com/eth-brownie/brownie)
interface and CLI
