+++ moonworm/generators/brownie_contract.py.template
