+++ moonworm/generators/cli.py.template
