+++ moonworm/generators/contract.py.template
